# docs/swagger.yaml
definitions:
...
  controller.ErrorResponse:
    properties:
      msg:
        type: string
    type: object
...
  dto.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      role:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
...
  model.ChangePassword:
    description: User password change data
    properties:
      confirm_password:
        example: newpassword123
        type: string
      current_password:
        example: oldpassword123
        type: string
      new_password:
        example: newpassword123
        minLength: 8
        type: string
    required:
    - confirm_password
    - current_password
    - new_password
    type: object
...
  model.RegisterUser:
    description: User registration data
    properties:
      email:
        example: john@example.com
        type: string
      full_name:
        example: John Doe
        type: string
      password:
        example: password123
        minLength: 8
        type: string
      role:
        enum:
        - user
        - admin
        example: user
        type: string
      username:
        example: johndoe
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - full_name
    - password
    - username
    type: object
...
  model.UpdateUser:
    description: User registration data
    properties:
      email:
        example: john@example.com
        type: string
      full_name:
        example: John Doe
        type: string
      role:
        enum:
        - user
        - admin
        example: user
        type: string
      username:
        example: johndoe
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - full_name
    - username
    type: object
...
info:
  contact: {}
  description: Fiber go web framework based REST API boilerplate
  title: Fiber Go API
  version: "1.0"
paths:
...
  /api/v1/register:
    post:
      consumes:
      - application/json
      description: Create a new user.
      parameters:
      - description: Create new user
        in: body
        name: register
        required: true
        schema:
          $ref: '#/definitions/model.RegisterUser'
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            $ref: '#/definitions/dto.User'
        "400":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "401":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "409":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: create a new user
      tags:
      - User
...
  /api/v1/users:
    get:
      consumes:
      - application/json
      description: Get all exists user.
      parameters:
      - description: Page no
        in: query
        name: page
        type: integer
      - description: records per page
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.User'
            type: array
        "400":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "401":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "403":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: get all exists user
      tags:
      - User
  /api/v1/users/{id}:
    delete:
      consumes:
      - application/json
      description: delete user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema: {}
        "401":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "403":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "404":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: delete a user
      tags:
      - User
    get:
      consumes:
      - application/json
      description: a user.
      parameters:
      - description: User ID (UUID format)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.User'
        "400":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "401":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "403":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "404":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: get a user
      tags:
      - User
    put:
      consumes:
      - application/json
      description: first_name, last_name, is_active, is_admin only
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Update a user
        in: body
        name: updateuser
        required: true
        schema:
          $ref: '#/definitions/model.UpdateUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.User'
        "400":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "401":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "403":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "404":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: update a user
      tags:
      - User
  /api/v1/users/{id}/change-password:
    post:
      consumes:
      - application/json
      description: Change user password with verification of current password
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Password change data
        in: body
        name: changePassword
        required: true
        schema:
          $ref: '#/definitions/model.ChangePassword'
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema: {}
        "400":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "401":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "403":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "404":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: change user password
      tags:
      - User
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"